pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "Simple agent sample using agent_framework OpenAI Assistants client"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  # Core framework includes OpenAI client
  "agent-framework-core>=1.0.0b251007",
]

[tool.uv]
package = true
dev-dependencies = ["ruff>=0.6.0", "poethepoet>=0.25"]

[project.scripts]
weather-agent = "weather_agent.main:main_entry"


[tool.uv.sources]
agent-framework-core = { path = "/home/reoring/dev/codroid/refs/agent-framework/python/packages/core", editable = true }

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.poe.tasks.lint]
shell = "ruff check ${PY_SRCS:-.} && ruff format --check ${PY_SRCS:-.}"

[tool.poe.tasks.format]
shell = "ruff format ${PY_SRCS:-.}"

[tool.poe.tasks.fix]
shell = "ruff check ${PY_SRCS:-.} --fix && ruff format ${PY_SRCS:-.}"
